Module Store.Name

type t

The name of an entity which may be watched. Note that not all these entities are stored in the xenstore tree (e.g. @introduceDomain)

val introduceDomain : t

The special watch when a domain is created

val releaseDomain : t

The special watch when a domain shuts down

val of_string : string -> t

val to_string : t -> string

val is_relative : t -> bool

is_relative t is true if t contains a relative path

val make_absolute : t -> string -> t

make_absolute t basename returns t as an absolute path, treating paths relative to basename

val to_key : t -> string list

to_key t returns a key which can be used as a path in a Trie