Module Store.Path

type t

Represents an absolute path within the xenstore tree

exception Doesnt_exist of string

val doesnt_exist : t -> 'a

doesnt_exist path raises the Doesnt_exist exception

val getdomainpath : int -> t

getdomainpath domid returns the default directory for domid

val create : string -> t -> t

create path default is the absolute path of path, where if path is relative then it is resolved relative to default

val to_name : t -> Name.t

to_name t returns the associated Name.t, suitable for watching

val to_string : t -> string

to_string t returns t as a '/'-separated path string

val to_string_list : t -> string list

to_string_list t returns t as a list of path element strings

val of_string_list : string list -> t

of_string_list l returns l as an instance of type t

val get_hierarchy : t -> t list

get_hierarchy t returns all t's on the path from the root node to t

val get_common_prefix : t -> t -> t

get_common_prefix a b returns the largest common prefix of a and b

val get_parent : t -> t

get_parent t returns the parent node of t. The parent of the root node is itself.

val make_relative : t -> Name.t -> Name.t

make_relative base name returns a Name.t which is name transformed into a relative path from base