Module Xenstore_server__Call

val reply : Xenstore_server.Store.t -> Xenstore_server.Connection.t -> Xs_protocol.t -> Xs_protocol.t

reply store con request modifies store according to the request in request from connection con and returns the response to be sent to the client.