Module Xs_protocol.Token

type t

A token is associated with every watch and returned in the callback.

val to_debug_string : t -> string

to_string token is a debug-printable version of token.

val to_user_string : t -> string

to_user_string token is the user-supplied part of token.

val of_string : string -> t

of_string str_rep is the token resulting from the unmarshalling of str_rep.

val to_string : t -> string

to_string token is the marshalled representation of token.