Module Xs_protocol

XenStore protocol.

type t

A valid packet.

module Op : sig ... end

module ACL : sig ... end

Access control lists.

module Parser : sig ... end

Incrementally parse packets.

module type IO = sig ... end

exception Unknown_xenstore_operation of int32

exception Response_parser_failed of string

type ('a, 'b) result =

| Ok of 'a

| Exception of 'b

module PacketStream : functor (IO : IO) -> sig ... end

val to_bytes : t -> bytes

val get_tid : t -> int32

val get_ty : t -> Op.t

val get_data : t -> string

val get_rid : t -> int32

val create : int32 -> int32 -> Op.t -> string -> t

module Token : sig ... end

module Response : sig ... end

module Request : sig ... end

module Unmarshal : sig ... end

exception Enoent of string

exception Eagain

exception Eexist

exception Invalid

exception Error of string

val response : string -> t -> t -> (t -> 'a option) -> 'a

response debug_hint sent received unmarshal returns the unmarshalled response corresponding to the received packet relative to the sent packet.

type address =

| Unix of string

| Domain of int

val string_of_address : address -> string

val domain_of_address : address -> int